pyproject.toml
[tool.poetry]
name = "feedback-report-generator"
version = "0.1.0"
description = ""
authors = ["Sachin Kumar <21f2000143@ds.study.iitm.ac.in>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.3"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
celery = "^5.4.0"
redis = "^5.2.0"
reportlab = "^4.2.5"
flower = "^2.0.1"
djangorestframework-simplejwt = "^5.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
